# AllAboard_swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: AllAboard API
host: 'localhost'
schemes:
  - http
...
paths:
...
  /station/station:
    post:
      tags:
        - station
      summary: Add a new station
      operationId: addStation
      consumes:
        - application/json
      produces: 
        - application/json
      parameters:
        - in: body
          name: station 
          description: Station object 
          required: true
          schema:
            $ref: '#/definitions/Station'
      responses:
        '200':
          description: Station created successfully
  /station/station/{stationId}:
    put:
      tags:
        - station
      summary: Update a station 
      operationId: updateStation
      parameters:
        - name: stationId
          in: path 
          description: ID of the station to update 
          required: true
          type: integer
        - name: station
          in: body
          description: Updated station object 
          required: true
          schema: 
            $ref: '#/definitions/Station'
      responses:
        200:
          description: Station updated successfully
        400: 
          description: Invalid request data
        404:
          description: Station not found
    
    delete: 
      tags:
        - station 
      summary: Delete a station 
      operationId: deleteStation
      parameters:
        - name: stationId
          in: path
          description: ID of the station to delete
          required: true
          type: integer
      responses:
        200:
          description: Station deleted successfully
        400:
          description: Invalid request data 
        404:
          description: Station not found  
...
  /train/train:
    post:
      tags:
        - train 
      summary: Add a new train 
      operationId: addTrain
      consumes: 
        - application/json
      produces: 
        - application/json 
      parameters: 
        - in: body 
          name: train 
          description: Train object 
          required: true 
          schema: 
            $ref: '#/definitions/Train'
      responses: 
        '200': 
          description: Train created successfully 
  /train/train/{trainId}:
    put:
      tags:
        - train
      summary: Update a train
      operationId: updateTrain
      parameters:
        - name: trainId
          in: path 
          description: ID of the train to update 
          required: true
          type: integer
        - name: train
          in: body
          description: Updated train 
          required: true
          schema: 
            $ref: '#/definitions/Train'
      responses:
        200:
          description: Train updated successfully
        400: 
          description: Invalid request data
        404:
          description: Train not found
    
    delete:
      tags:
        - train 
      summary: Delete a train 
      operationId: deleteTrain
      parameters:
        - name: trainId
          in: path
          description: ID of the train to delete
          required: true
          type: integer
      responses:
        200:
          description: Train deleted successfully
        400:
          description: Invalid request data 
        404:
          description: Train not found  
  /train/getTrainsBasedOnStation:
    get:
      tags:
        - train
      summary: Get trains based on station
      operationId: getTrainsBasedOnStation
      parameters:
        - name: from
          in: query
          description: The source station
          required: true
          type: string
        - name: to
          in: query
          description: The destination station
          required: true
          type: string
        - name: date
          in: query
          description: The travel date
          required: true
          type: string
          format: date
      produces:
        - application/json
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Train'
...
definitions:
...
  Station:
    type: object
    properties:
      station_id:
        type: integer
      station_name:
        type: string
...
  TrainClass:
    type: object 
    properties:
      train_class_type:
        type: string 
      capacity:
        type: integer
      base_price: 
        type: number 
  TrainSeat: 
    type: object 
    properties: 
      seat_type: 
        type: string 
      class_type:
        type: string
      seat_number:
        type: string
      is_booked: 
        type: boolean 
      seat_price:
        type: number
  Train:
    type: object
    properties:
      train_id:
        type: integer
      train_name:
        type: string
      source_station:
        type: string
      destination_station:
        type: string
      travel_date:
        type: string
        format: date
      departure_time:
        type: string
        format: time
      train_classes:
        type: array
        items:
          $ref: '#/definitions/TrainClass'
      train_seats: 
        type: array 
        items: 
          $ref: '#/definitions/TrainSeat'
...